requirements.txt
requests
python-ics
